+++ lib/roar/representer/xml.rb
@@ -9,7 +9,6 @@ module Roar
     module XML
       def self.included(base)
         base.class_eval do
-          include Representer
           include Representable::XML
 
           extend ClassMethods
@@ -18,11 +17,6 @@ module Roar
       end
 
       module InstanceMethods
-        def to_node(*args)
-          before_serialize(*args)
-          super
-        end
-
         # Generic entry-point for rendering.
         def serialize(*args)
           to_xml(*args)
@@ -45,7 +39,6 @@ module Roar
             :class        => Feature::Hypermedia::Hyperlink,
             :extend       => XML::HyperlinkRepresenter,
             :exec_context => :decorator,
-            :getter         => lambda { |*| links.values }, # links is LinkCollection, we just render a list of Hyperlinks.
             } # TODO: merge with JSON.
         end